+++ ElementX/Sources/Screens/KnockRequestsListScreen/View/KnockRequestsListScreen.swift
@@ -78,11 +78,11 @@ struct KnockRequestsListScreen: View {
 struct KnockRequestsListScreen_Previews: PreviewProvider, TestablePreview {
     static let emptyViewModel = KnockRequestsListScreenViewModel.mockWithInitialState(.init(requests: []))
     
-    static let viewModel = KnockRequestsListScreenViewModel.mockWithInitialState(.init(requests: [.init(id: "1", userID: "@alice:matrix.org", displayName: "Alice", avatarURL: nil, timestamp: "Now", reason: "Hello"),
+    static let viewModel = KnockRequestsListScreenViewModel.mockWithInitialState(.init(requests: [.init(eventID: "1", userID: "@alice:matrix.org", displayName: "Alice", avatarURL: nil, timestamp: "Now", reason: "Hello"),
                                                                                                   // swiftlint:disable:next line_length
-                                                                                                  .init(id: "2", userID: "@bob:matrix.org", displayName: "Bob", avatarURL: nil, timestamp: "Now", reason: "Hello this one is a very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very long reason"),
-                                                                                                  .init(id: "3", userID: "@charlie:matrix.org", displayName: "Charlie", avatarURL: nil, timestamp: "Now", reason: nil),
-                                                                                                  .init(id: "4", userID: "@dan:matrix.org", displayName: "Dan", avatarURL: nil, timestamp: "Now", reason: "Hello! It's a me! Dan!")]))
+                                                                                                  .init(eventID: "2", userID: "@bob:matrix.org", displayName: "Bob", avatarURL: nil, timestamp: "Now", reason: "Hello this one is a very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very very long reason"),
+                                                                                                  .init(eventID: "3", userID: "@charlie:matrix.org", displayName: "Charlie", avatarURL: nil, timestamp: "Now", reason: nil),
+                                                                                                  .init(eventID: "4", userID: "@dan:matrix.org", displayName: "Dan", avatarURL: nil, timestamp: "Now", reason: "Hello! It's a me! Dan!")]))
                                                                                       
     static var previews: some View {
         NavigationStack {
