+++ ElementX/Sources/Services/Room/RequestToJoinProxy.swift
@@ -8,26 +8,6 @@
 import Foundation
 import MatrixRustSDK
 
-enum RequestToJoinProxyError: Error {
-    case sdkError(Error)
-}
-
-// sourcery: AutoMockable
-protocol RequestToJoinProxyProtocol {
-    var eventID: String { get }
-    var userID: String { get }
-    var displayName: String? { get }
-    var avatarURL: URL? { get }
-    var reason: String? { get }
-    var formattedTimestamp: String? { get }
-    var isSeen: Bool { get }
-    
-    func accept() async -> Result<Void, RequestToJoinProxyError>
-    func decline() async -> Result<Void, RequestToJoinProxyError>
-    func ban() async -> Result<Void, RequestToJoinProxyError>
-    func markAsSeen() async -> Result<Void, RequestToJoinProxyError>
-}
-
 struct RequestToJoinProxy: RequestToJoinProxyProtocol {
     private let requestToJoin: RequestToJoin
